Migrate bevy-tutti to flat TuttiEngine bundle
Replaces the god-object Arc<TuttiEngine> (wrapped as TuttiEngineResource)
with per-subsystem Bevy resources destructured from the upstream flat
bundle: TuttiGraphRes, TuttiDriverRes, TransportRes, MeteringRes,
MidiBusRes, MidiIoRes, SamplerRes, SoundFontRes, AnalysisRes, NeuralRes,
and an AudioConfig resource carrying sample_rate + channels.

Graph-edit systems now take ResMut<TuttiGraphRes>, mutate via the new
typed TuttiGraph API (add / master / pipe_output / node::<T> /
node_mut::<T>), and call graph.commit() once per system after a batch
of edits. TransportHandle / MeteringHandle / MidiBus / AnalysisHandle
and Arc<Sampler> / Arc<SoundFontSystem> / Arc<NeuralEngine> flow
through their own Res<...> parameters.

TuttiDriver is wrapped in a Mutex because cpal's Stream is Send but not
Sync; driver ops (restart / set_device) are infrequent and exclusive.

Plugin editor attach now passes a ThreadLockedRawWindowHandleWrapper
(obtained from RawHandleWrapper::get_handle on the main thread) to
PluginHandle::open_editor, matching the new HasWindowHandle-based
signature. MIDI note dispatch for MidiSequence uses MidiBus::queue on
MidiUnitId with MidiEvent::note_on / note_off constructors. MPE is
reduced to a compile-time stub until the bundle exposes an observer
hook; live analysis enable/disable is similarly stubbed because
AnalysisHandle has no runtime toggle in this version of tutti.

The TuttiEngineResource newtype, inject_tokio_handle_system, and the
tutti umbrella re-exports that no longer exist (MidiHandle,
register_plugin*, tutti::Channel/Note at the root, NeuralEngine, etc.)
are removed. Umbrella re-exports now route through the new tutti
subsystem modules (tutti::midi, tutti::sampler, tutti::synth,
tutti::export, tutti::automation, tutti::neural, tutti::plugin,
tutti::dsp_nodes, tutti::core).

+++ src/native_window.rs
@@ -13,7 +13,7 @@ pub fn native_view_ptr(raw: raw_window_handle::RawWindowHandle) -> Option<u64> {
     match raw {
         RawWindowHandle::AppKit(h) => Some(h.ns_view.as_ptr() as u64),
         RawWindowHandle::Win32(h) => Some(isize::from(h.hwnd) as u64),
-        RawWindowHandle::Xlib(h) => Some(h.window as u64),
+        RawWindowHandle::Xlib(h) => Some(h.window),
         RawWindowHandle::Xcb(h) => Some(h.window.get() as u64),
         RawWindowHandle::Wayland(h) => Some(h.surface.as_ptr() as u64),
         _ => None,
